code cleaning and compiler optimizations
