--- OctomapServer.h
@@ -1,154 +1,0 @@
-/*
- * Copyright (c) 2010-2013, A. Hornung, University of Freiburg
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the University of Freiburg nor the names of its
- *       contributors may be used to endorse or promote products derived from
- *       this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-#ifndef OCTOMAPSERVER_H
-#define OCTOMAPSERVER_H
-
-#include <chrono>
-
-#include <pcl/point_types.h>
-#include <pcl/conversions.h>
-#include <pcl/common/transforms.h>
-#include <pcl/sample_consensus/method_types.h>
-#include <pcl/sample_consensus/model_types.h>
-#include <pcl/segmentation/sac_segmentation.h>
-#include <pcl/io/pcd_io.h>
-#include <pcl/filters/extract_indices.h>
-#include <pcl/filters/passthrough.h>
-
-#include <octomap/octomap.h>
-#include <octomap/OcTreeKey.h>
-
-#include "debug_definitions.h"
-//#define COLOR_OCTOMAP_SERVER // turned off here, turned on identical ColorOctomapServer.h - easier maintenance, only maintain OctomapServer and then copy and paste to ColorOctomapServer and change define. There are prettier ways to do this, but this works for now
-
-#ifdef COLOR_OCTOMAP_SERVER
-#include <octomap/ColorOcTree.h>
-#endif
-
-class OctomapServer {
-
-public:
-#ifdef COLOR_OCTOMAP_SERVER
-  typedef pcl::PointXYZRGB PCLPoint;
-  typedef pcl::PointCloud<pcl::PointXYZRGB> PCLPointCloud;
-  typedef octomap::ColorOcTree OcTreeT;
-#else
-  typedef pcl::PointXYZ PCLPoint;
-  typedef pcl::PointCloud<pcl::PointXYZ> PCLPointCloud;
-  typedef octomap::OcTree OcTreeT;
-#endif
-
-  OctomapServer();
-  virtual ~OctomapServer();
-
-  virtual void insertCloudCallback(PCLPointCloud::Ptr pc, const Eigen::Matrix4f& sensorToWorld, const Eigen::Matrix4f& sensorToBase, const Eigen::Matrix4f& baseToWorld);
-
-  OcTreeT* m_octree;
-protected:
-  inline static void updateMinKey(const octomap::OcTreeKey& in, octomap::OcTreeKey& min) {
-    for (unsigned i = 0; i < 3; ++i)
-      min[i] = std::min(in[i], min[i]);
-  };
-
-  inline static void updateMaxKey(const octomap::OcTreeKey& in, octomap::OcTreeKey& max) {
-    for (unsigned i = 0; i < 3; ++i)
-      max[i] = std::max(in[i], max[i]);
-  };
-
-  /// Test if key is within update area of map (2D, ignores height)
-  inline bool isInUpdateBBX(const OcTreeT::iterator& it) const {
-    // 2^(tree_depth-depth) voxels wide:
-    unsigned voxelWidth = (1 << (m_maxTreeDepth - it.getDepth()));
-    octomap::OcTreeKey key = it.getIndexKey(); // lower corner of voxel
-    return (key[0] + voxelWidth >= m_updateBBXMin[0]
-            && key[1] + voxelWidth >= m_updateBBXMin[1]
-            && key[0] <= m_updateBBXMax[0]
-            && key[1] <= m_updateBBXMax[1]);
-  }
-
-  /**
-  * @brief update occupancy map with a scan labeled as ground and nonground.
-  * The scans should be in the global map frame.
-  *
-  * @param sensorOrigin origin of the measurements for raycasting
-  * @param ground scan endpoints on the ground plane (only clear space)
-  * @param nonground all other endpoints (clear up to occupied endpoint)
-  */
-  virtual void insertScan(const Eigen::Vector3f& sensorOriginTf, const PCLPointCloud& ground, const PCLPointCloud& nonground);
-
-  /// label the input cloud "pc" into ground and nonground. Should be in the robot's fixed frame (not world!)
-  void filterGroundPlane(const PCLPointCloud& pc, PCLPointCloud& ground, PCLPointCloud& nonground) const;
-
-  octomap::KeyRay m_keyRay;  // temp storage for ray casting
-  octomap::OcTreeKey m_updateBBXMin;
-  octomap::OcTreeKey m_updateBBXMax;
-
-  double m_maxRange;
-  std::string m_worldFrameId; // the map frame
-  std::string m_baseFrameId; // base of the robot for ground plane filtering
-  bool m_useHeightMap;
-
-  double m_res;
-  unsigned m_treeDepth;
-  unsigned m_maxTreeDepth;
-
-  double m_pointcloudMinX;
-  double m_pointcloudMaxX;
-  double m_pointcloudMinY;
-  double m_pointcloudMaxY;
-  double m_pointcloudMinZ;
-  double m_pointcloudMaxZ;
-  double m_occupancyMinZ;
-  double m_occupancyMaxZ;
-  // double m_minSizeX;
-  // double m_minSizeY;
-  bool m_filterSpeckles;
-
-  bool m_filterGroundPlane;
-  double m_groundFilterDistance;
-  double m_groundFilterAngle;
-  double m_groundFilterPlaneDistance;
-
-  bool m_compressMap;
-
-  bool m_initConfig;
-
-  // downprojected 2D map:
-  bool m_incrementalUpdate;
-  // nav_msgs::OccupancyGrid m_gridmap;
-  bool m_publish2DMap;
-  bool m_mapOriginChanged;
-  octomap::OcTreeKey m_paddedMinKey;
-  unsigned m_multires2DScale;
-  bool m_projectCompleteMap;
-  bool m_useColoredMap;
-};
-
-#endif
