+++ include/OctomapServer.h
@@ -1,0 +1,93 @@
+/*
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef OCTOMAPSERVER_H
+#define OCTOMAPSERVER_H
+
+#include <chrono>
+
+#include <pcl/point_types.h>
+#include <pcl/conversions.h>
+#include <pcl/common/transforms.h>
+#include <pcl/sample_consensus/method_types.h>
+#include <pcl/sample_consensus/model_types.h>
+#include <pcl/segmentation/sac_segmentation.h>
+#include <pcl/io/pcd_io.h>
+#include <pcl/filters/extract_indices.h>
+#include <pcl/filters/passthrough.h>
+
+#include <octomap/octomap.h>
+#include <octomap/OcTreeKey.h>
+
+#include "debug_definitions.h"
+//#define COLOR_OCTOMAP_SERVER // turned off here, turned on identical ColorOctomapServer.h - easier maintenance, only maintain OctomapServer and then copy and paste to ColorOctomapServer and change define. There are prettier ways to do this, but this works for now
+
+#ifdef COLOR_OCTOMAP_SERVER
+#include <octomap/ColorOcTree.h>
+#endif
+
+class OctomapServer {
+
+public:
+#ifdef COLOR_OCTOMAP_SERVER
+  typedef pcl::PointXYZRGB PCLPoint;
+  typedef pcl::PointCloud<pcl::PointXYZRGB> PCLPointCloud;
+  typedef octomap::ColorOcTree OcTreeT;
+#else
+  typedef pcl::PointXYZ PCLPoint;
+  typedef pcl::PointCloud<pcl::PointXYZ> PCLPointCloud;
+  typedef octomap::OcTree OcTreeT;
+#endif
+
+  OctomapServer();
+  virtual ~OctomapServer();
+
+  virtual void insertCloudCallback(PCLPointCloud::Ptr pc, const Eigen::Matrix4f& sensorToWorld);
+
+  OcTreeT* m_octree;
+protected:
+  inline static void updateMinKey(const octomap::OcTreeKey& in, octomap::OcTreeKey& min) {
+    for (unsigned i = 0; i < 3; ++i)
+      min[i] = std::min(in[i], min[i]);
+  };
+
+  inline static void updateMaxKey(const octomap::OcTreeKey& in, octomap::OcTreeKey& max) {
+    for (unsigned i = 0; i < 3; ++i)
+      max[i] = std::max(in[i], max[i]);
+  };
+
+  /**
+  * @brief update occupancy map with a scan.
+  * The scans should be in the global map frame.
+  *
+  * @param sensorOrigin origin of the measurements for raycasting
+  * @param nonground all other endpoints (clear up to occupied endpoint)
+  */
+  virtual void insertScan(const Eigen::Vector3f& sensorOriginTf, const PCLPointCloud& cloud);
+  
+  octomap::KeyRay m_keyRay;  // temp storage for ray casting
+  octomap::OcTreeKey m_updateBBXMin;
+  octomap::OcTreeKey m_updateBBXMax;
+
+  double m_maxRange;
+
+  double m_res;
+  unsigned m_treeDepth;
+  unsigned m_maxTreeDepth;
+
+  bool m_compressMap;
+};
+
+#endif
