+++ include/debug_definitions.h
@@ -1,3 +1,18 @@
+/*
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #define RED "\x1B[31m"
 #define GRN "\x1B[32m"
 #define YEL "\x1B[33m"
@@ -19,7 +34,7 @@
     std::cout << __FUNCTION__ << ": Line " << __LINE__ << ": " << RED << str \
               << RESET << std::endl;                                         \
   } while (false)
-#define Dbg(str)                                                           \
+#define Dbg(str)                                                            \
   do {                                                                       \
     std::cout << __FUNCTION__ << ": Line " << __LINE__ << ": " << BLUE << str \
               << RESET << std::endl;                                         \
